+++ vstgui/lib/platform/win32/direct2d/d2ddrawcontext.h
@@ -62,8 +62,8 @@ public:
 	// CDrawContext
 	void moveTo (const CPoint &point);
 	void lineTo (const CPoint &point);
-	void drawLines (const CPoint* points, const long& numberOfLines);
-	void drawPolygon (const CPoint *pPoints, long numberOfPoints, const CDrawStyle drawStyle = kDrawStroked);
+	void drawLines (const CPoint* points, const int32_t& numberOfLines);
+	void drawPolygon (const CPoint *pPoints, int32_t numberOfPoints, const CDrawStyle drawStyle = kDrawStroked);
 	void drawRect (const CRect &rect, const CDrawStyle drawStyle = kDrawStroked);
 	void drawArc (const CRect &rect, const float startAngle1, const float endAngle2, const CDrawStyle drawStyle = kDrawStroked);
 	void drawEllipse (const CRect &rect, const CDrawStyle drawStyle = kDrawStroked);
